backend/pyproject.toml
[tool.poetry]
name = "semantico-backend"
version = "0.1.0"
description = ""
authors = ["Anis Koubaa <mohamed.koubaa@kit.edu>"]
readme = "README.md"
packages = [{include = "semantico_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
pymupdf = "^1.22.5"
fastapi = {extras = ["all"], version = "^0.104.1"}
requests = "^2.31.0"
pyzotero = "^1.5.18"
pdfminer-six = "^20221105"
numpy = "^1.26.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
